requirements.txt
pyautogui==0.9.54
Pillow==10.4.0
adbutils==2.8.0